refactor: separate command group drag and drop components
