+++ cmd/gomander/frontend/src/components/modals/CommandGroup/common/CommandGroupCommandsField/CommandGroupCommandsField.tsx
@@ -4,113 +4,35 @@ import {
   type DragEndEvent,
   type DragOverEvent,
   PointerSensor,
-  useDroppable,
   useSensor,
   useSensors,
 } from "@dnd-kit/core";
 import {
   arrayMove,
   SortableContext,
-  useSortable,
   verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
 import { ChevronRight, X } from "lucide-react";
 import { useMemo } from "react";
-import React from "react";
 import { useFormContext } from "react-hook-form";
 
+import {
+  ADDED_COMMANDS,
+  AVAILABLE_COMMANDS,
+} from "@/components/modals/CommandGroup/common/CommandGroupCommandsField/constants.ts";
+import { DraggableCommandItem } from "@/components/modals/CommandGroup/common/CommandGroupCommandsField/DraggableCommandItem.tsx";
+import { DroppableContainer } from "@/components/modals/CommandGroup/common/CommandGroupCommandsField/DroppableContainer.tsx";
 import type { FormSchemaType } from "@/components/modals/CommandGroup/common/formSchema.ts";
 import {
   FormDescription,
   FormField,
   FormItem,
-  FormLabel,
   FormMessage,
 } from "@/components/ui/form.tsx";
-import type { Command } from "@/contracts/types.ts";
 import { isDefined } from "@/helpers/mapHelpers.ts";
 import { useCommandStore } from "@/store/commandStore.ts";
 
-const AVAILABLE_COMMANDS = "available-commands";
-const ADDED_COMMANDS = "added-commands";
-
-const DraggableCommandItem = ({
-  command,
-  rightComponent,
-}: {
-  command: Command;
-  rightComponent?: React.ReactNode;
-}) => {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id: command.id });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-    opacity: isDragging ? 0.5 : 1,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className="group flex items-center p-1.5 pl-3 bg-card hover:bg-accent border rounded"
-    >
-      <div {...attributes} {...listeners} className="flex-1 cursor-grab">
-        <span className="text-sm">{command.name}</span>
-      </div>
-
-      {rightComponent}
-    </div>
-  );
-};
-
-const DroppableContainer = ({
-  variant,
-  children,
-  className,
-}: {
-  variant: typeof AVAILABLE_COMMANDS | typeof ADDED_COMMANDS;
-  children: React.ReactNode;
-  className?: string;
-}) => {
-  const { setNodeRef } = useDroppable({ id: variant });
-
-  return (
-    <div className={className}>
-      {variant === AVAILABLE_COMMANDS && (
-        <h4 className="font-medium text-sm mb-2">Available Commands</h4>
-      )}
-      {variant === ADDED_COMMANDS && (
-        <FormLabel className="font-medium text-sm mb-2">
-          Group Commands
-        </FormLabel>
-      )}
-      <div
-        ref={setNodeRef}
-        className="h-80 max-h-[calc(100vh-400px)] p-3 border rounded-lg overflow-y-auto overflow-x-hidden"
-      >
-        <div className="space-y-2">{children}</div>
-        {React.Children.count(children) === 0 && (
-          <div className="flex items-center justify-center h-full text-muted-foreground text-sm text-center">
-            {variant === AVAILABLE_COMMANDS
-              ? "No more commands available"
-              : "Drag commands here"}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-export const NewCommandGroupCommandsField = () => {
+export const CommandGroupCommandsField = () => {
   const allCommands = useCommandStore((state) => state.commands);
   const form = useFormContext<FormSchemaType>();
