+++ cmd/gomander/frontend/src/components/modals/CommandGroup/common/CommandGroupCommandsField/constants.ts
@@ -1,0 +1,2 @@
+export const AVAILABLE_COMMANDS = "available-commands";
+export const ADDED_COMMANDS = "added-commands";
